requirements.txt
tornado==2.4.1
mysql-python
redis==2.7.2
sqlalchemy
pyOpenSSL